face/requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0